backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
exceptiongroup==1.3.1
fastapi==0.129.0
filelock==3.21.2
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==26.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyPDF2==3.0.1
PyYAML==6.0.3
regex==2026.1.15
rich==14.3.2
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.2
shellingham==1.5.4
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.1.0
typer==0.23.0
typer-slim==0.23.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
watchdog==6.0.0
supabase==2.28.0
python-dotenv==1.2.1
